feature/19 test: ControllerTest 공통 기능 분리
 - CommonControllerTest를 상속받는 구조로 변경
